rename metastructure to meta_structure
