+++ bim2sim/kernel/aggregation.py
@@ -10,9 +10,9 @@ import numpy as np
 from bim2sim.decision import BoolDecision, DecisionBunch
 from bim2sim.decorators import cached_property
 from bim2sim.kernel import attribute
-from bim2sim.metastructure.__init__ import ProductBased
-from bim2sim.metastructure import bps, hvac
-from bim2sim.metastructure.hvac import HVACPort
+from bim2sim.meta_structure.__init__ import ProductBased
+from bim2sim.meta_structure import bps, hvac
+from bim2sim.meta_structure.hvac import HVACPort
 from bim2sim.kernel.hvac.hvac_graph import HvacGraph
 from bim2sim.kernel.units import ureg
 from bim2sim.utilities.common_functions import filter_instances
