+++ bim2sim/meta_structure/__init__.py
