+++ bim2sim/meta_structure/hvac.py
@@ -12,11 +12,11 @@ import numpy as np
 from bim2sim.decision import ListDecision, DecisionBunch
 from bim2sim.decorators import cached_property
 from bim2sim.kernel import condition, attribute
-from bim2sim.metastructure import Port, ProductBased
+from bim2sim.meta_structure import Port, ProductBased
 from bim2sim.kernel.ifc2python import get_ports as ifc2py_get_ports
 from bim2sim.kernel.ifc2python import get_predefined_type
 from bim2sim.kernel.units import ureg
-from bim2sim.metastructure import IFCBased
+from bim2sim.meta_structure import IFCBased
 
 logger = logging.getLogger(__name__)
 quality_logger = logging.getLogger('bim2sim.QualityReport')
