+++ bim2sim/plugins/PluginAixLib/bim2sim_aixlib/models/__init__.py
@@ -1,7 +1,7 @@
 ﻿"""Package for Python representations of HKESim models"""
 import bim2sim.kernel.aggregation as aggregation
 from bim2sim.export import modelica
-from bim2sim.metastructure import hvac
+from bim2sim.meta_structure import hvac
 from bim2sim.kernel.units import ureg
