+++ bim2sim/plugins/PluginEnergyPlus/bim2sim_energyplus/task/ep_add_2b_bounds.py
@@ -15,7 +15,7 @@ from OCC.Core.Extrema import Extrema_ExtFlag_MIN
 from OCC.Core.TopoDS import TopoDS_Face
 from OCC.Core.gp import gp_Pnt
 
-from bim2sim.metastructure.bps import SpaceBoundary2B, ThermalZone, Door, \
+from bim2sim.meta_structure.bps import SpaceBoundary2B, ThermalZone, Door, \
     Window
 from bim2sim.task.base import ITask
 from bim2sim.utilities.common_functions import get_spaces_with_bounds
