+++ bim2sim/task/bps/verify.py
@@ -1,5 +1,5 @@
-from bim2sim.metastructure.__init__ import Material
-from bim2sim.metastructure.bps import BPSProductWithLayers, LayerSet, Layer
+from bim2sim.meta_structure.__init__ import Material
+from bim2sim.meta_structure.bps import BPSProductWithLayers, LayerSet, Layer
 from bim2sim.kernel.units import ureg
 from bim2sim.task.base import ITask
 from bim2sim.utilities.common_functions import all_subclasses, filter_instances
