+++ bim2sim/task/hvac/hvac.py
@@ -16,8 +16,8 @@ from bim2sim.kernel.aggregation import Consumer, \
     ConsumerHeatingDistributorModule, GeneratorOneFluid
 from bim2sim.kernel.aggregation import PipeStrand, UnderfloorHeating, \
     ParallelPump
-from bim2sim.metastructure.__init__ import ProductBased, Port, Material
-from bim2sim.metastructure import hvac
+from bim2sim.meta_structure.__init__ import ProductBased, Port, Material
+from bim2sim.meta_structure import hvac
 from bim2sim.kernel.hvac import hvac_graph
 from bim2sim.kernel.hvac.hvac_graph import HvacGraph
 from bim2sim.task.base import ITask, Playground
