+++ test/unit/kernel/aggregation/test_parallelpumps.py
@@ -3,7 +3,7 @@
 import networkx as nx
 
 from bim2sim.kernel import aggregation
-from bim2sim.metastructure import hvac
+from bim2sim.meta_structure import hvac
 from bim2sim.kernel.hvac.hvac_graph import HvacGraph
 from test.unit.kernel.helper import SetupHelperHVAC
