+++ test/unit/kernel/aggregation/test_pipestrand.py
@@ -1,7 +1,7 @@
 import unittest
 
 from bim2sim.kernel import aggregation
-from bim2sim.metastructure import hvac
+from bim2sim.meta_structure import hvac
 from bim2sim.kernel.hvac.hvac_graph import HvacGraph
 from bim2sim.kernel.units import ureg
 from test.unit.kernel.helper import SetupHelperHVAC
